requirements.txt
langchain>=0.1.14
langchain-community>=0.0.27
langchain-google-genai>=0.0.9
google-generativeai>=0.3.2
Flask>=2.3.2
requests>=2.31.0
python-dotenv>=1.0.1
schedule>=1.2.1